requirements.txt
docker
kafka-python
msgpack
typeguard
temporenc

pytest
mypy
flake8